Refactoring to base class ready for moving Interceptor
